==> notebooks/02_Data_Cleaning.ipynb <==
# Data Cleaning  
**Objective:**   
- Address data quality issues identified during EDA.  
  
- Handle missing values, outliers, and incorrect data types.  
   
- Ensure data integrity for modeling.  

### Import Libraries and Load Data

import pandas as pd
import numpy as np

df = pd.read_csv('../data/raw/cs-training.csv', index_col=0)

### Handle Missing Values

**NumberOfDependents:**

# Handle the missing value of NumberOfDependents via imputation
df['NumberOfDependents'] = df['NumberOfDependents'].fillna(0)

**MonthlyIncome:**

# Include Flag for missing MonthlyIncome
df['MonthlyIncomeMissing'] = df['MonthlyIncome'].isna().astype(int)

# Handle the missing value of MonthlyIncome via imputation
df['MonthlyIncome'] = df['MonthlyIncome'].fillna(0)

### Correcting Data Types

**NumberOfDependents:**

df['NumberOfDependents'] = df['NumberOfDependents'].astype('int64')

### Address Data Inconsistencies

**Age:**

# Remove ages less than 0
df = df[df['age'] > 0] 

**RevolvingUtilizationOfUnsecuredLines:**

# Cap values to 10 to address extreme outliers
df['RevolvingUtilizationOfUnsecuredLines'] = df['RevolvingUtilizationOfUnsecuredLines'].clip(upper=10)

**DebtRatio:**

# Calculate the 99th percentile
debt_ratio_99perc = df['DebtRatio'].quantile(0.99)
print(f'99th percentile of DebtRatio: {debt_ratio_99perc}')

# Cap values at the 99th percentile
df['DebtRatio'] = df['DebtRatio'].clip(upper=debt_ratio_99perc)

### Remove Duplicates

df.drop_duplicates(inplace=True)

### Save Cleaned Data

df.to_csv('../data/processed/cleaned_data.csv', index=False)
